requirements-conda.txt
pytorch==0.4.1
torchvision
pandas